+++ src/Core/Components/Dialog/FluentDialogProvider.razor.cs
@@ -15,7 +15,7 @@ public partial class FluentDialogProvider : IDisposable
     private NavigationManager NavigationManager { get; set; } = default!;
 
     /// <summary>
-    /// Constructs an instance of <see cref="FluentToastContainer"/>.
+    /// Constructs an instance of <see cref="FluentToastProvider"/>.
     /// </summary>
     public FluentDialogProvider()
     {