+++ src/Core/Components/Toast/Services/CountdownTimer.cs
@@ -12,7 +12,7 @@ internal class CountdownTimer : IDisposable
     internal CountdownTimer(int timeout, CancellationToken cancellationToken = default)
     {
         _ticksToTimeout = 100;
-        _timer = new PeriodicTimer(TimeSpan.FromMilliseconds(timeout * 10));
+        _timer = new PeriodicTimer(TimeSpan.FromMilliseconds(timeout/100));
         _cancellationToken = cancellationToken;
     }