+++ src/Core/Components/Toast/Services/InternalToastContext.cs
@@ -4,11 +4,11 @@ internal sealed class InternalToastContext
 {
     public Dictionary<string, FluentToast> References { get; set; } = new();
 
-    public FluentToastContainer ToastContainer { get; }
+    public FluentToastProvider ToastProvider { get; }
 
-    public InternalToastContext(FluentToastContainer container)
+    public InternalToastContext(FluentToastProvider provider)
     {
-        ToastContainer = container;
+        ToastProvider = provider;
     }
 
     internal void Register(FluentToast toast)